Made code more modular for adding enemies
Further split the BasicPathfinding script and EnemyBehaviour scripts into 2

+++ Assets/Scripts/Player/PlayerStatManager.cs
@@ -4,15 +4,15 @@ using UnityEngine;
 
 public class PlayerStatManager : MonoBehaviour
 {
-    public int maxHealth = 100;
-    public int currentHealth;
+    public float maxHealth = 100;
+    public float currentHealth;
 
     private void Start()
     {
         currentHealth = maxHealth;    
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currentHealth -= damage;
         Debug.Log("Took damage: Health is now " + currentHealth);
